pyproject.toml
[project]
name = "homeaccmt26"
version = "0.1.0"
description = "Multi-tenant double-entry accounting application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlmodel>=0.0.22",
  "pyjwt>=2.9.0",
  "pwdlib[argon2]>=0.2.0",
  "python-dotenv>=1.0.0",
  "alembic>=1.14.0",
  "pydantic-settings>=2.6.0",
  "aiosqlite>=0.20.0",
  "email-validator>=2.3.0",
  "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.0",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
